+++ src/entities/adventurer/states.ts
@@ -2,12 +2,9 @@ import { adventurerStand } from './states/stand';
 import { adventurerStandDraw } from './states/stand-draw';
 import { adventurerStandHold } from './states/stand-hold';
 import { adventurerStandShoot } from './states/stand-shoot';
-import { adventurerCrouch } from './states/crouch';
-import { adventurerCrouchLeft } from './states/crouch-left';
-import { adventurerCrouchRight } from './states/crouch-right';
 import { adventurerRunRight } from './states/run-right';
 import { adventurerRunLeft } from './states/run-left';
-import { adventurerRoll } from './states/slide';
+import { adventurerRoll } from './states/roll';
 import { adventurerJumpPrep } from './states/jump-prep';
 import { adventurerJump} from './states/jump';
 import { adventurerJumpLeft } from './states/jump-left';
@@ -18,6 +15,7 @@ import { adventurerFallRight } from './states/fall-right';
 import { adventurerAirDraw } from './states/air-draw';
 import { adventurerAirHold } from './states/air-hold';
 import { adventurerAirShoot } from './states/air-shoot';
+import { adventurerCrouch } from './states/crouch';
 
 export const adventurerStates: PhiniteStateMachine.States.State<Phecs.Entity>[] = [
   adventurerStand,
@@ -26,8 +24,9 @@ export const adventurerStates: PhiniteStateMachine.States.State<Phecs.Entity>[] 
   adventurerStandShoot,
 
   adventurerCrouch,
-  adventurerCrouchLeft,
-  adventurerCrouchRight,
+  // adventurerCrouch,
+  // adventurerCrouchLeft,
+  // adventurerCrouchRight,
 
   adventurerRunRight,
   adventurerRunLeft,
