--- src/entities/adventurer/states/crouch-left.ts
@@ -1,18 +1,0 @@
-import { baseCrouch } from './base-crouch';
-import { TransitionType } from '../../../lib/phinite-state-machine/transition-type';
-import { StateMerge } from '../../../lib/phinite-state-machine/state-merge';
-import { SpriteComponent } from '../../../components/sprite-component';
-
-export const adventurerCrouchLeft: PhiniteStateMachine.States.State<Phecs.Entity> = StateMerge(baseCrouch, {
-  id: 'adventurer-crouch-left',
-  onEnter(entity: Phecs.Entity) {
-    entity.getComponent(SpriteComponent).sprite.flipX = true;
-  },
-  transitions: [
-    {
-      type: TransitionType.PressControl,
-      control: 'right',
-      to: 'adventurer-crouch-right'
-    }
-  ]
-});
