+++ src/entities/adventurer/test-prefab.ts
@@ -1,15 +1,22 @@
+import { SpriteComponent } from "../../components/sprite-component";
+import { BoundsComponent } from "../../components/bounds-component";
+import { PhysicsBodyComponent } from "../../components/physics-body-component";
+
 export const adventurerTestPrefab = {
-  name: 'adventurer',
-
-  tags: "sprite,physics-body,attachment,hurtbox",
-
-  boundsKey: "adventurer-bounds",
-
-  texture: "adventurer-core",
-  frame: 0,
-  maxVelocityX: 350,
-
-  hurtboxesKey: "adventurer-hurtboxes",
-
-  attachmentDebug: false,
-}
+  components: [
+    {
+      component: SpriteComponent,
+      data: {
+        texture: "adventurer-core",
+        frame: 0,
+      }
+    },
+    {
+      component: BoundsComponent,
+      data: {
+        boundsKey: "adventurer-bounds",
+      }
+    },
+    PhysicsBodyComponent
+  ]
+};
