+++ src/scenes/prefab-test-scene.ts
@@ -23,36 +23,26 @@ export class PrefabTestScene extends BaseScene {
   }
 
   create() {
-    // this.phecs.phEntities.registerPrefab('knight', knightPrefab);
+    this.phecs.phEntities.registerPrefab('adventurer', adventurerTestPrefab);
 
-    /*
     this.phecs.phSystems.registerSystems(
       [
         HasBoundsSystem
       ]
     );
-    */
 
-    const barrel = this.add.sprite(300, 200, 'adventurer-core');
-    barrel.anims.play('adventurer-roll')
 
-    this.input.on('pointerdown', () => barrel.anims.play('adventurer-roll'))
-
-    // this.stateRegistrar.registerSets([
-      // { id: 'arrow', states: arrowStates },
-    // ]);
-
-    /*
-    let frameIndex = 22;
+    let frameIndex = 21;
     const frameText = this.add.text(300, 50, `Frame ${frameIndex}`);
 
-    const knight = this.phecs.phEntities.createPrefab('knight', {}, 1, 150, 200);
+    const entity = this.phecs.phEntities.createPrefab('adventurer', {}, 1, 150, 200);
 
-    knight.getComponent(SpriteComponent).sprite.setScale(3);
-    knight.getComponent(SpriteComponent).sprite.setFrame(frameIndex);
+    entity.getComponent(SpriteComponent).sprite.setScale(4);
+    entity.getComponent(SpriteComponent).sprite.setFrame(frameIndex);
 
-    knight.getComponent(PhysicsBodyComponent).body.allowGravity = false;
+    entity.getComponent(PhysicsBodyComponent).body.allowGravity = false;
 
+    /*
     knight.getComponent(SpriteComponent).sprite.anims.play('knight-attack');
 
     this.input.keyboard.on('keydown', (e: any) => {
@@ -68,12 +58,10 @@ export class PrefabTestScene extends BaseScene {
       knight.getComponent(SpriteComponent).sprite.setFrame(frameIndex);
       frameText.setText(`Frame ${frameIndex}`);
     });
+    */
 
     this.phecs.start();
 
     this.cameras.main.setBackgroundColor('#777777');
-
-    this.knight = knight;
-    */
   }
 }
